refactor(toolbar): split out toolbar into it's own file
